Publish distribution files to PyPI
Add steps to publish distribution files to PyPI when a tag is created.

We also split the step of building the package distribution as we don't
need to do that for every Python version, so there is no much use to
have it inside a matrix.

This also parallelize jobs as much as possible, so protolint, tests and
building the distribution packages are done in parallel.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>