ci: Create GitHub Release automatically
Add a GitHub job to create the GitHub Release automatically when
a version is tagged. Also use the GitHub feature to create the CHANGELOG
automatically. For this the CHANGELOG.md is removed in favor of
a RELEASE_NOTES.md files that should only contain an explanation on the
release main changes.

Also add the release notes for the current release.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>