Create contributing guide
All the content of the README.md file that refers to development is
moved to the CONTRIBUTING.md guide, as the README.md should be more
targeted to users rather than developers (as it is used for the PyPI
page too).

Add also the steps to create a release to the contributing guide.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>