Add grid-connection-point metadata
This commit introduces a `grid.Metadata` message, which contains the item
`rated_fuse_current`. This is the rating of the fuse at the grid connection
point.
This rating specifies the maximum amount of current, measured in amperes,
that can flow in or out of each of the 3 phases individually.
The current _i_ A at the grid connection point must comply with the
following constraint: : `-rated_fuse_current <= i <= rated_fuse_current`

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>