Introduce exclusion bounds
In the messages `common.Metric` and `common.MetricAggregation`,
`system_bounds` has now been replaced by `system_exclusion_bounds` and
`system_inclusion_bounds`. A metric's `value` now has to comply with the
following constraints:

* `value <= system_exclusion_bounds.lower` OR
  `system_exclusion_bounds.upper <= value`

* `system_inclusion_bounds.lower <= value <= system_inclusion_bounds.upper`

`system_inclusion_bounds` behave in the same manner as the earlier
`system_bounds`.

The following diagram illustrates the relationship between the exclusion
and inclusion bounds.
```
  inclusion.lower                              inclusion.upper
<-------|============|------------------|============|--------->
               exclusion.lower    exclusion.upper
```
`----` values here are disallowed and wil be rejected, and
`====` values here are allowed and will be accepted.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>