Support solar and battery DC links for inverters
The `Inverter` message has been updated to support hybrid and solar
inverters. The `dc` field has been renamed to `dc_battery`, and a new field
`dc_solar` has been added.

The `dc_battery` field is used to report the DC electricity flowing to/from
the linked battery, and is applicable to battery and hybrid inverters.

The `dc_solar` field is used to report the DC electricity flowing to/from
the linked solar panels, and is applicable to solar and hybrid inverters.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>