Use common proto API's pypi package in python lib
The pypi package `frequenz-api-common` is being added as a dependency to
the python package definition, instead of generating the proto definitions
using `protoc`. This is required, otherwise each proto library depending on
`frequenz-api-common` will generate its own python modules for
`frequenz-api-common`, resulting in multiple definition of the common data
structures.

Note that this forces the minimum required python version of the
`frequenz-api-microgrid` package to be 3.11, as a transitive dependency
inherited from the `frequenz-api-common` package.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>