Use common proto API's pypi package in python lib
Python files for proto definitions in `frequenz-api-common` are no longer
being generated using protoc. Instead, the pypi package
`frequenz-api-common` is being added as a dependency to the python package
definition.
This is required, otherwise each proto library depending on
`frequenz-api-common` will generate its own python modules for
`frequenz-api-common`, resulting in multiple definition of the same common
structures.

Note that this forces the minimum required python version to be 3.11, as a
transitive dependency from the `frequenz-api-common` package.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>