Rename submodules to include the path
Even names without the path make more sense, they are a bit useless (and
because of that a bit error prone) as all `git submodule` commands take
the path and not the name.

Also when configuring other tools like *dependabot* it is not always
that clear which one should be used (spoiler alert, it seems to be the
path too, not the name).

So it is better to just remove the ambiguity.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>