Add support for relays
Relays are electromagnetic switches that control circuit breakers in the
microgrid, e.g., to connect or disconnect an inverter from the grid.

This commit allows clients to expect data for such relays in a microgrid,
and invoke the following commands:
  - `StartComponent`
  - `StopComponent`

Note that the relay.Relay message does not have `properties` or `data`.
`state` and `errors` should be enough to capture the information we need.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>