Add support for precharge modules
Precharging a DC (Direct Current) bus refers to a controlled process in
electrical systems, e.g., when connecting an inverter to a battery, where
the voltage across a DC bus is gradually increased from a low level to a
desired level before connecting a load or initiating normal operation.
This process is primarily used to prevent sudden surges of current that
can occur when a load is directly connected to a high-voltage DC source.

This commits adds support for interacting with precharge modules to the
Microgrid API. Clients can expect data from such components, and can
invoke the following RPCs:
  - `StartComponent`
  - `StopComponent`

Note that the precharge_module.PrechargeModule message does not have
`properties` or `data`. `state` and `errors` should be enough to capture
the information we need.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>