Add voltage transformer component
This commit adds the voltage transformer component to the microgrid.

Voltage transformers are used to step up or step down the voltage,
keeping the power somewhat constant by increasing or decreasing the
current. If voltage is stepped up, current is stepped down, and vice
versa. Note that voltage transformers have efficiency losses, so the
output power is always less than the input power.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>