Rename RPC parameters to be more consistent with the RPC names
The RPC parameters were named in a way that was not consistent with the
RPC names. This commit renames the RPC parameters to be more consistent
with the RPC names. The following rules have been followed, in order:
1. If an RPC is named `GetX`, then its request message is named
   `GetXRequest` and its response message is named `X`. Same for
   `SubscribeX`.
2. If an RPC is named `VerbX`, then its request message is named
   `VerbXRequest`. Its response message is named `VerbXResponse`, unless
   the response type is `google.protobuf.Empty`.

Also, the documentation in the proto files has been updated to be more
consistent with each other.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>