Add support for filtering metrics in data streams
This commit adds support for filtering metrics in data streams. The
request messages for receiving data streams have now been extended to
consist of a list of metrics to be streamed. This allows the user to
request only the metrics they are interested in, instead of receiving
all of them. If this list is empty, then no data will be streamed, and
the service will return an error.

The filter for the data stream RPCs are optional. This  is implemented
by putting the metrics field into a nested message, which is then
added as a field in the request message. This makes the filter
optional, so users can apply filters only when they intend to. This
allows the filter to be omitted, which will result in the server
returning all metrics for the given component or sensor.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>