Return a valid-until timestamp from RPC `SetComponentPowerActive`
This commit allows the RPC `SetComponentPowerActive` to return a
timestamp until which the given power command will stay in effect. After
this timestamp, the component's active power will be set to 0, if the API
receives no further command to change it before then.

Typically, this timestamp will be set to the current time plus 60 seconds.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>