Bump the api-common dependency to v0.5.4
The new v0.5.4 release of the api-common repository is now backwards
compatible with the old v0.3.x releases. We bump the dependency so
microgrid API can be used together with other APIs that depend on the
newer common API version.

Please note that the googleapis-common-protos dependency is also bumped
to v1.56.4, which is the version that api-common v0.5.4 depends on.

Strictly speaking, this is a breaking change too, as projects might need
to bump their googleapis-common-protos dependency to v1.56.4 too if they
specify it explicitly, but this is highly unlikely and very easy to fix.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>