Allow clients to specify validity durations for bounds
This commit adds a field named `validity_duration` to the
`AddComponentBoundsRequest` message, which can be used by clients to
specify a custom default duration to bounds while adding them. If not
specified, a default validity duration of 5 seconds is applied.

The decision to use an enumerated list of allowed durations has been
impacted by our intention to not allow users to set arbitrarily long
validity durations on bounds. Using an enum makes it clear while using the
API.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>