Update microgrid.proto to use ElectricalComponent terminology
The term `Component` has been changes in the common API to
`ElectricalComponent`. This commit introduces this new terminology in
the Microgrid API as well, replacing all instances of `Component` with
`ElectricalComponent`. This change is made to maintain consistency
across the API and to clarify that the components being referred to are
electrical in nature.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>