Add `ramp_rate` to SetComponentPowerActive/ReactiveRequest
This commit adds a `ramp_rate` field to the
`SetComponentPowerActiveRequest` and `SetComponentPowerReactiveRequest`
messages. This field allows users to specify the desired rate of change
for the power output of electrical components in a microgrid. This is
useful in scenarios where the rate of change of power is critical tio
the use-case of the client.

The `ramp_rate` field is an optional unsigned integer that represents
the magnitude of the ramp in watts per second (for active power) or
volt-amperes per second (for reactive power). The service will apply
the ramp in the correct direction (up or down) to reach the target
power. The actual rate of change will be capped by the component's
configured maximum ramp rate, ensuring that the system operates within
the component's limits.

If the `request_lifetime` ends before the ramp is complete, the
component will be set to its standby state. If a new command arrives
mid-ramp, the current ramp will be interrupted, and a new one will
begin according to the new command's parameters. If the `ramp_rate` is
not provided, the component will use its configured maximum ramp rate.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>