Stream responses for `SetElectricalComponentPowerActive/Reactive` RPCs
The response messages for the `SetElectricalComponentPowerActive` and
`SetElectricalComponentPowerReactive` RPCs have been changed to return a
stream of responses instead of a single response. This allows the server
to provide ongoing status updates for the request, which is useful when
the electrical component cannot immediately set its output power to the
requested value. The initial response will indicate whether the request
was accepted or rejected, and a subsequent response will indicate
whether the request was successful, failed, or overridden by another
request. If the request was accepted and successful, the response will
also contain the timestamp until which the given power command will stay
in effect, after which the electrical component will be returned to its
default state.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>