Revert validity_duration of `AddComponentBoundsRequest` to 5 seconds
The previous validity duration of 60 seconds was too long for the
microgrid use case, where a shorter duration of 5 seconds is more
appropriate. This is because these bounds restraint the capabilities
the electrical components, and the way for clients to revert their
bounds is to stop sending bounds. This is the desired approach, since it
keeps the overhead on the client side low. Given this style of
intteraction, the older 5s default is more appropriate, since it allows
clients to quickly revert their bounds without having to wait for a long
time.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>